[refactor] #4 Shorten 기능 데이터 전송 방법 수정 및 이름변경
요청 url 수정
repository 함수명 변경
서비스 인터페이스 삭제
