+++ src/main/java/me/nexters/chop/domain/url/Url.java
@@ -22,7 +22,7 @@ public class Url extends BaseTime {
     @Column(nullable = false)
     private String longUrl;
 
-    @Column(nullable = false)
+    @Column(nullable = false, unique = true)
     private String shortUrl;
 
     @Builder
